split log config into mods
